Dynamic particle pool: auto-size from device dimensions, remove template
Replace compile-time ParticlePool<N> template with dynamically-allocated
pool sized in begin() from particleDensity() * numLeds. Each generator
provides its density fraction (Fire 0.75, Water 0.50, Lightning 0.67)
and the pool scales automatically to any device — no hardcoded caps.

32x32 display now gets 768 fire particles (was capped at 64). An 89-LED
hat gets 66 (was also 64). No per-device configuration needed.

Changes:
- ParticlePool: remove template, add begin(capacity) with heap alloc
- ParticleGenerator: remove template, add virtual particleDensity()
- Fire/Water/Lightning: inherit ParticleGenerator (not <N>), override
  particleDensity(), scaledMaxParticles() returns pool_.getCapacity()
- All uint8_t maxParts -> uint16_t (pools can exceed 255)
- Flash size drops ~2KB (eliminated 3 template instantiations)

+++ blinky-things/generators/Fire.cpp
@@ -178,7 +178,7 @@ void Fire::spawnParticles(float dt) {
     }
 
     // Spawn sparks using layout-aware spawn region with variety
-    uint8_t maxParts = scaledMaxParticles();
+    uint16_t maxParts = scaledMaxParticles();
     for (uint8_t i = 0; i < sparkCount && pool_.getActiveCount() < maxParts; i++) {
         float x, y;
         getSpawnPosition(x, y);
